举一反三修复整数溢出问题
# message auto-generated for no-merge-commit merge:
merge timeline_device_feat_****** into master

【********】【开发自提】【安全排查】【MindStudio Insight】【TOPN】【数据溢出】【举一反三】返回值赋值会存在整数溢出问题

Created-by: insightDeveloper

Match-id-d98d0f830fde76f835835689ccde1f5ea7e34f5f

+++ server/src/modules/advisor/protocol/AdvisorProtocolResponse.h
@@ -27,7 +27,7 @@ struct AffinityOptimizerData {
 };
 
 struct AffinityOptimizerResBody {
-    uint32_t size{};
+    uint64_t size{};
     std::string dbPath;
     std::vector<AffinityOptimizerData> datas;
 };
@@ -46,7 +46,7 @@ struct AffinityAPIData {
 };
 
 struct AffinityAPIResBody {
-    uint32_t size{};
+    uint64_t size{};
     std::string dbPath;
     std::vector<AffinityAPIData> datas;
 };
@@ -65,7 +65,7 @@ struct OperatorFusionData {
 };
 
 struct OperatorFusionResBody {
-    uint32_t size{};
+    uint64_t size{};
     std::string dbPath;
     std::vector<OperatorFusionData> datas;
 };
@@ -82,7 +82,7 @@ struct AICpuOperatorData {
 };
 
 struct AICpuOperatorResBody {
-    uint32_t size{};
+    uint64_t size{};
     std::string dbPath;
     std::vector<AICpuOperatorData> datas;
 };
@@ -99,7 +99,7 @@ struct AclnnOperatorData {
 };
 
 struct AclnnOperatorResBody {
-    uint32_t size{};
+    uint64_t size{};
     std::string dbPath;
     std::vector<AclnnOperatorData> datas;
 };
@@ -116,7 +116,7 @@ struct OperatorDispatchData {
 };
 
 struct OperatorDispatchResBody {
-    uint32_t size{};
+    uint64_t size{};
     std::string dbPath;
     std::vector<OperatorDispatchData> data;
 };
